backend/app/services/geo_scorer.py
"""
GEO Scorer - Geographic Optimization (Phase 2 Stub)
Placeholder for multi-location and geographic optimization scoring
Will be fully implemented in Week 5 (Phase 2)
"""

import logging
from typing import Dict, Any, List, Optional

logger = logging.getLogger(__name__)


class GEOScorer:
    """
    Geographic Optimization Scorer - Phase 2 Feature

    Future Scoring breakdown (45 points total):
    - Multi-location optimization: 15 points
    - Local citations consistency: 10 points
    - Geographic content relevance: 10 points
    - Google Business Profile integration: 10 points

    Current Status: STUB - Returns 0 points
    Implementation Timeline: Week 5 (Phase 2)
    """

    def calculate_geo_score(self, site_data: Dict[str, Any]) -> Dict[str, Any]:
        """
        Calculate geographic optimization score

        Args:
            site_data: Site data for GEO analysis

        Returns:
            GEO score analysis (currently returns 0 - stub)
        """
        logger.info("GEO Scorer called - Phase 2 feature (stub)")

        return {
            "geo_score": 0,
            "max_score": 45,
            "percentage": 0,
            "status": "Phase 2 - Not Yet Implemented",
            "planned_features": [
                "Multi-location page optimization",
                "Local citations consistency check",
                "Geographic keyword analysis",
                "Google Business Profile integration",
                "Service area targeting",
                "Local schema markup validation"
            ],
            "breakdown": {
                "multi_location": {
                    "score": 0,
                    "max": 15,
                    "status": "Coming in Phase 2"
                },
                "local_citations": {
                    "score": 0,
                    "max": 10,
                    "status": "Coming in Phase 2"
                },
                "geographic_content": {
                    "score": 0,
                    "max": 10,
                    "status": "Coming in Phase 2"
                },
                "gbp_integration": {
                    "score": 0,
                    "max": 10,
                    "status": "Coming in Phase 2"
                }
            },
            "recommendations": [
                "GEO scoring will be available in Phase 2 (Week 5)",
                "Focus on AEO optimization in the meantime",
                "Prepare by ensuring NAP consistency and LocalBusiness schema"
            ],
            "implementation_timeline": "Week 5 - Phase 2",
            "message": "Geographic optimization scoring is a planned feature for Phase 2. Current focus: AEO (Answer Engine Optimization)."
        }

    def get_geo_readiness(self, site_data: Dict[str, Any]) -> Dict[str, Any]:
        """
        Check if site is ready for GEO scoring (Phase 2 prep)

        Args:
            site_data: Site data to analyze

        Returns:
            Readiness assessment for future GEO implementation
        """
        # Basic readiness checks we can do now
        html_content = site_data.get('html', '')
        has_local_business = False

        # Check for LocalBusiness schema (indicates local business)
        if html_content:
            from bs4 import BeautifulSoup
            import json

            soup = BeautifulSoup(html_content, 'lxml')
            jsonld_scripts = soup.find_all('script', type='application/ld+json')

            for script in jsonld_scripts:
                try:
                    schema_data = json.loads(script.string)
                    schemas_to_check = schema_data if isinstance(schema_data, list) else [schema_data]

                    for schema in schemas_to_check:
                        if isinstance(schema, dict):
                            schema_type = schema.get('@type', '')
                            if schema_type == 'LocalBusiness':
                                has_local_business = True
                                break
                except:
                    continue

        return {
            "ready_for_geo": has_local_business,
            "has_local_business_schema": has_local_business,
            "recommended_actions": [
                "Add LocalBusiness schema if you serve specific geographic areas",
                "Ensure NAP (Name, Address, Phone) consistency",
                "Create location-specific pages if you have multiple locations",
                "Claim and optimize Google Business Profile"
            ] if not has_local_business else [
                "Good! You have LocalBusiness schema",
                "GEO scoring will analyze this in Phase 2",
                "Focus on AEO optimization in the meantime"
            ],
            "phase_2_features": [
                "Multi-location optimization analysis",
                "Citation consistency checking across directories",
                "Geographic content relevance scoring",
                "Google Business Profile integration and insights"
            ]
        }

    def validate_location_data(self, location_data: Dict[str, Any]) -> Dict[str, Any]:
        """
        Validate manual location data input (for future use)

        Args:
            location_data: Manual location information

        Returns:
            Validation results
        """
        # Placeholder for future manual location data validation
        required_fields = ['name', 'address', 'city', 'state', 'zip', 'phone']
        missing_fields = []

        for field in required_fields:
            if field not in location_data or not location_data[field]:
                missing_fields.append(field)

        is_valid = len(missing_fields) == 0

        return {
            "valid": is_valid,
            "missing_fields": missing_fields,
            "message": "Location data validated" if is_valid else f"Missing required fields: {', '.join(missing_fields)}",
            "note": "GEO scoring is a Phase 2 feature - this validates data structure only"
        }


# Future implementation notes for Phase 2 (Week 5):
"""
PHASE 2 GEO IMPLEMENTATION ROADMAP

Week 5 Implementation Plan:

1. Multi-Location Optimization (15 points):
   - Detect location-specific pages
   - Validate unique content per location
   - Check for location-specific schema
   - Analyze internal linking between locations
   - Score: 15/15 if 3+ locations with unique content

2. Local Citations Consistency (10 points):
   - Integrate with citation data sources
   - Check NAP consistency across directories
   - Validate citation count and quality
   - Score: 10/10 if NAP consistent across 10+ citations

3. Geographic Content Relevance (10 points):
   - Analyze geographic keywords
   - Check service area mentions
   - Validate location-specific content depth
   - Score: 10/10 if strong geo-content strategy

4. Google Business Profile Integration (10 points):
   - API integration with GBP
   - Check profile completeness
   - Analyze review count and ratings
   - Validate posts and updates frequency
   - Score: 10/10 if optimized GBP

Data Sources to Integrate:
- Google Business Profile API
- Citation data (Moz Local, BrightLocal, or custom crawler)
- Google Maps API for location validation
- Review aggregation APIs

Manual Input Features:
- Location management interface
- Citation tracking
- Review monitoring
- Multi-location dashboard

Dependencies to Add:
- google-api-python-client (for GBP integration)
- geopy (for address validation and geocoding)
- Additional APIs based on final feature set
"""
